Proposal for v0.1.0 release
In an effort to enable an official 'offline' state via the web service, I decided it would be cleaner to just support more dynamically configured states.  But because I want to avoid causing upgrade issues for our small existing user base, i geared all the changes so that upon upgrade eerything should still behave exactly the same.

In the long run we'd like to possibly change the returned HTTP response code to be a 200 for every state, thus showing that the server is working.  This decision is based upon the ugliness of trying to associate HTTP response codes to our states.  At that point we would be relying on the returned message for state checking.

In the configuration file you will see that the [permanants] (we named that poorly) section has been replaced with a section per state, and a new section called [priorities]. Under priorities you would provide a comma separated list, from lowest priority to highest.  Lowest priority should equate to your default state, typically called 'enabled'.  This list is also how the config parser looks up sections in the configuration.  If you have a section defined for a state, but it is not in the priorities.states list it will be ignored.